// frontend/src/api/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: `${string}://${string}` | (string & {});
};

export type ProblemDetails = {
  type?: null | string;
  title?: null | string;
  status?: null | number;
  detail?: null | string;
  instance?: null | string;
};

export type StationResponse = {
  code: string;
  name: string;
  latitude: number;
  longitude: number;
};

export type StationsResponse = {
  stations: Array<StationResponse>;
};

export type GetStationsData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/v1/stations";
};

export type GetStationsErrors = {
  /**
   * Bad Gateway
   */
  502: ProblemDetails;
};

export type GetStationsError = GetStationsErrors[keyof GetStationsErrors];

export type GetStationsResponses = {
  /**
   * OK
   */
  200: StationsResponse;
};

export type GetStationsResponse = GetStationsResponses[keyof GetStationsResponses];